+++ spec/rubocop/cop/rails/delegate_spec.rb
@@ -93,6 +93,14 @@ RSpec.describe RuboCop::Cop::Rails::Delegate do
     RUBY
   end
 
+  it 'ignores the method in the body with arguments' do
+    expect_no_offenses(<<-RUBY.strip_indent)
+      def fox
+        bar(42).fox
+      end
+    RUBY
+  end
+
   it 'ignores private delegations' do
     expect_no_offenses(<<-RUBY.strip_indent)
         private def fox # leading spaces are on purpose